requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# LangChain (Open Source)
langchain==0.1.13
langchain-community==0.0.30
langchain-core==0.1.37
langchain-text-splitters==0.0.1

# Vector Stores (Optional)
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Tools
ollama==0.1.9
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
ccxt==4.2.56

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.26.0
psutil==5.9.6
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.11.0
